requirements.txt
medmnist>=2.2.3
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0

# Task 2
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.41.0

# Task 3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
open-clip-torch>=2.23.0
